Clock speed specific builds
Previously I was configuring all bootloader files as 20MHz. The reason I
could do this is F_CPU is used only for determining the number of loops
required to generate a 100us delay(see
https://github.com/zevero/avr_boot/issues/1#issuecomment-153698038). My
theory was that this was the minimum delay length and a longer delay
wouldn't be a problem so by setting the value of F_CPU at 2000000 the
delay was 100us with a 20MHz clock and greater with a slower clock. Now
that the boards system allows the bootloader file name to be
automatically determined, adding clock speed specific builds comes only
at the cost of a longer build time, larger makeall.bat, and larger
installation size, none of which are terribly significant. If this can
allow faster reads from the SD card then I think it's worth it. This
change will be required anyway once serial upload support is added to
avr_boot.