Change MENU_ID to match Arduino AVR Boards
There is a bug(https://github.com/arduino/Arduino/issues/5260) in the
Arduino IDE that causes the menu item text to not update from the
previously selected board if the menu title text is the same but the
MENU_ID is different. Arduino AVR Boards defines menu.cpu=Processor so I
must do the same to reduce the chances of a user encountering this
issue. Note that this issue still may be caused by other boards
packages.