Add externalprogrammers.txt
This file is necessary for compatibility with the Eclipse Arduino plugin
which will not show programmers for any other package.

Due to removeing the toolsDependencies items, avr_boot is currently
incompatible with the plugin but it appears the plugin developer may
have added compatibility for this configuration.

Programmers entries from Arduino AVR Boards 1.6.13.